Squashed commit of the following:
commit 9db6d74cc16308faa963a942553873785a86dd6e
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 21:20:54 2024 -0400

    Update Overworld Game Scene.unity

commit 5a2aca20600b2e2aaa2af5f697f1df1f721279b0
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 21:19:46 2024 -0400

    blastvfx script fix

commit 6c2a29c304427c054b292492476939c548802dce
Merge: d1a342e 26df46f
Author: Raeus Aranguren-Viegas <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 21:14:55 2024 -0400

    Merge pull request #83 from Raaee/blastVFX

    blast VFX complete

commit 26df46f1ca683b52088df443764147928a7421e0
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 21:14:36 2024 -0400

    blast VFX complete

commit d1a342e48c01908bb39376262e6ee6a1a8c4b316
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 18:31:04 2024 -0400

    fixed tank animation not using correct controller

commit f98a95bf024bfa083477a633222655851b2eccca
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 04:26:38 2024 -0400

    char select menu fix

commit b9e8e3e573c4d8d0a6656fb2d3c1283887c2718e
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 04:16:10 2024 -0400

    balance stuff. changed crystal spawners and enemy damages

commit c297dd65c29981869a42a24208b917527370663a
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 03:50:58 2024 -0400

    balancing and changed mana type

commit cc79cded94122c985fcae4ea0f4c5e7b0468db99
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 03:07:14 2024 -0400

    pink mage damage change and enemy hp

commit 4cc7bb25bbb2602e4c464a5748f8f453d95affa0
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 03:01:03 2024 -0400

    card upgrade % and deleted projMan

commit 73a1ca34ad2c410d42083d528a75573b1e6ffbf7
Merge: c54f554 be9053e
Author: Raeus Aranguren-Viegas <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 02:51:13 2024 -0400

    Merge pull request #82 from Raaee/TeleportFlash

    Teleport flash

commit be9053eb6aae4dc5b643dc94340bafe54151d17c
Merge: 69be6d7 c54f554
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 02:50:31 2024 -0400

    Merge branch 'main' into TeleportFlash

commit c54f5548aaf5585ad7c74de5a0f47f31adf78ade
Merge: c203c0a 66cd4a9
Author: Raeus Aranguren-Viegas <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 02:46:27 2024 -0400

    Merge pull request #81 from Raaee/CrystalVFX-3.0

    projectile damage fix and player stat balancing, started crystal blast logic, fixed berserk functionality

commit 66cd4a9f98e68b6da0c9b0855f3ea15dbbf8db70
Merge: 3b49cc7 c203c0a
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 02:45:02 2024 -0400

    Merge branch 'main' into CrystalVFX-3.0

commit 3b49cc7fc724acf6f7939ac34c504379feb08281
Author: Raaee <99899402+Raaee@users.noreply.github.com>
Date:   Thu Apr 11 02:42:22 2024 -0400

    fixed projectile damage issue, started player stats balancing

commit ad2a772b605e149809d451f356e435f24a271332
Author: Isaacl78 <isaclugo2014@gmail.com>
Date:   Wed Apr 10 22:27:58 2024 -0400

    Bug fix

commit ab5629833e4d3ee6a2093ef344693c7451e85196
Author: Isaacl78 <isaclugo2014@gmail.com>
Date:   Wed Apr 10 22:27:43 2024 -0400

    Bezeker

commit 6ce188c8179d11771934deb7e9bc1fca631f5000
Author: Isaacl78 <isaclugo2014@gmail.com>
Date:   Wed Apr 10 17:44:55 2024 -0400

    Bux some bugs, try to work on projectile.

commit 69be6d78e3e932805b2b3d0193fdc14e872c9474
Author: dieg0rtega <160441988+dieg0rtega@users.noreply.github.com>
Date:   Wed Apr 10 17:00:44 2024 -0400

    Telport Visual added and Damage Visual improved

commit d0e740a97fbc5d748f7449e4c9f3f81520ec6041
Merge: ceba221 694b525
Author: Isaacl78 <isaclugo2014@gmail.com>
Date:   Wed Apr 10 15:05:30 2024 -0400

    Merge branch 'main' into CrystalVFX-3.0

commit ceba22146fb1679094123dec0e1c21470e4ce8f5
Author: Isaacl78 <isaclugo2014@gmail.com>
Date:   Tue Apr 9 23:51:44 2024 -0400

    Stated with the blast Attack

+++ Assets/Scripts/UI/UIManager.cs
@@ -89,8 +89,8 @@ public class UIManager : MonoBehaviour  {
         hpText.text = PlayerManager.Instance.hp.GetCurrentHP() + " / " + PlayerManager.Instance.hp.GetMaxHealth();
     }
     public void UpdateMana() {
-        manaFilling.fillAmount = (float) PlayerManager.Instance.mp.currentMP / PlayerManager.Instance.mp.maxMP;
-        manaText.text = PlayerManager.Instance.mp.currentMP + " / " + PlayerManager.Instance.mp.maxMP;
+        manaFilling.fillAmount = PlayerManager.Instance.mp.currentMP / PlayerManager.Instance.mp.maxMP;
+        manaText.text = (int) PlayerManager.Instance.mp.currentMP + " / " + (int) PlayerManager.Instance.mp.maxMP;
     }
     // Initial Set For The Crystal Bar
     public void ActivateCrystalBars(){
